+++ src/main/java/insane96mcp/survivalreimagined/module/world/item/FirestarterItem.java
@@ -1,21 +1,22 @@
 package insane96mcp.survivalreimagined.module.world.item;
 
+import insane96mcp.survivalreimagined.module.mining.feature.MiningCharge;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.stats.Stats;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.FlintAndSteelItem;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.UseAnim;
-import net.minecraft.world.item.Vanishable;
+import net.minecraft.world.item.*;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
 
@@ -44,7 +45,24 @@ public class FirestarterItem extends FlintAndSteelItem implements Vanishable {
         if (!(entity instanceof Player player))
             return stack;
         BlockHitResult blockHitResult = getPlayerPOVHitResult(player.level, player, ClipContext.Fluid.NONE);
-        super.useOn(new UseOnContext(player, player.getUsedItemHand(), blockHitResult));
+        BlockPos pos = blockHitResult.getBlockPos();
+        BlockState state = level.getBlockState(pos);
+        if (state.is(Blocks.TNT) || state.is(MiningCharge.MINING_CHARGE.block().get())) {
+            if (state.is(Blocks.TNT))
+                Blocks.TNT.onCaughtFire(state, level, pos, blockHitResult.getDirection(), player);
+            else
+                MiningCharge.MINING_CHARGE.block().get().onCaughtFire(state, level, pos, blockHitResult.getDirection(), player);
+            level.setBlock(pos, Blocks.AIR.defaultBlockState(), 11);
+            Item item = stack.getItem();
+            if (!player.isCreative()) {
+                stack.hurtAndBreak(1, player, (player1) -> player1.broadcastBreakEvent(player1.getUsedItemHand()));
+            }
+
+            player.awardStat(Stats.ITEM_USED.get(item));
+        }
+        else {
+            super.useOn(new UseOnContext(player, player.getUsedItemHand(), blockHitResult));
+        }
         player.getCooldowns().addCooldown(stack.getItem(), 20);
         return stack;
     }