+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Golden shield now protects 100 damage instead of giving a speed boost
+* Firestarter now properly lit up TNT and Mining charges
 
 ## 0.12.0
 * Added Fletching Table